paquo: restructure module
keep module structure more simple
